Refactor/Fix: locking issues no longer occur when multiple threads attempt to update the search index
It turns out I was using Lucene in a very inefficient way, and much like many other 'expensive' classes, an IndexWriter should be a static instance which is then shared/reused throughout the app. In addition, the IndexSearcher should use an IndexReader it gets from that same static instance, so the index can be queried safely while updates are being applied.

Thanks to this guy for his answer:

https://stackoverflow.com/questions/35053251/updating-lucene-index-frequently-causing-performance-degrade

So now, LuceneSearchService is injected as a singleton and is just a facade which calls the static LuceneSearchFunctions overloads.

Restructuring like this means that only a single folder lock is ever required on the index, and the writer itself co-ordinates thread-safety and concurrency (including that of any readers using the same index).

It also seems to have improved search performance quite dramatically, which is nice.